use size_t instead of int for array indices that can go beyond int range

+++ preproc/lib/preproc.cpp
@@ -43,7 +43,7 @@ static void normalize_intensity(int num_frames, int width, int height, float ***
 
 static void copy1d_to_3d(int num_frames, int width, int height, float *in, float ***out)
 {
-    int n = 0;
+    size_t n = 0;
     for(int k = 0; k < num_frames; k++)
     for(int j = 0; j < height; j++)
     for(int i = 0; i < width; i++)
@@ -54,7 +54,7 @@ static void copy1d_to_3d(int num_frames, int width, int height, float *in, float
 
 static void copy3d_to_1d(int num_frames, int width, int height, float ***in, float *out)
 {
-    int n = 0;
+    size_t n = 0;
     for(int k = 0; k < num_frames; k++)
     for(int j = 0; j < height; j++)
     for(int i = 0; i < width; i++)
